Updating IRON_FISH_YEAR_IN_BLOCKS due to block time change
