+++ ironfish/src/strategy.test.slow.ts
@@ -282,26 +282,3 @@ describe('Demonstrate the Sapling API', () => {
     })
   })
 })
-
-describe('Miners reward', () => {
-  let strategy: Strategy
-
-  beforeAll(() => {
-    strategy = new Strategy(new WorkerPool())
-  })
-
-  // see https://ironfish.network/docs/whitepaper/4_mining#include-the-miner-reward-based-on-coin-emission-schedule
-  // for more details
-  it('miners reward is properly calculated for year 0-1', () => {
-    let minersReward = strategy.miningReward(1)
-    expect(minersReward).toBe(5 * 10 ** 8)
-
-    minersReward = strategy.miningReward(100000)
-    expect(minersReward).toBe(5 * 10 ** 8)
-  })
-
-  it('miners reward is properly calculated for year 1-2', () => {
-    const minersReward = strategy.miningReward(2100001)
-    expect(minersReward).toBe(475614712)
-  })
-})
